+++ src/app/options/components/BadgeList.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react'
 import { Button } from '@/components/ui/Button'
-import { GitHubLogoIcon } from '@radix-ui/react-icons'
+import { Github } from 'lucide-react'
 import Link from '@/components/Link'
 
 const BadgeList: FC = () => {
@@ -8,7 +8,7 @@ const BadgeList: FC = () => {
     <div className="mx-auto mt-6 flex w-fit pb-8">
       <Button asChild size="lg" variant="ghost" className="rounded-full px-3 text-xl font-semibold text-primary">
         <Link href="https://github.com/tbdavid2019/chrome-webtalk">
-          <GitHubLogoIcon className="mr-1 size-6"></GitHubLogoIcon>
+          <Github className="mr-1 size-6"></Github>
           GitHub / 開源專案
         </Link>
       </Button>